+++ docs/user_guide/changelog.rst
@@ -11,7 +11,9 @@ New features
 - Ability to send different values for a weight tensor to different replica groups (see :numref:`grouping_tensor_weights`)
 - Support for ``tensor.index_fill_``
 - Support for ``torch.count_nonzero``
+- Support for ``torch.scatter_reduce``
 - Support for Mk2x.
+
 
 
 API changes